[user-input-manager] cleanup
